Added 'add' and 'preview' functionality
Each selection now has a customized list of items that can be added
under it. These items are also added in the proper position relative to
the other children.

The Preview tab now updates when you select a new TreeItem, and if
anything other than the Dictionary itself is selected, it will only
show the relevant Dictionary Entry.